+++ api/models/Examinee.js
@@ -1,0 +1,6 @@
+const mongoose = require('mongoose');
+const { User } = require('./User');
+
+const Examinee = User.discriminator('Examinee', new mongoose.Schema({}));
+
+module.exports.Examinee = Examinee;