+++ api/routes/users.js
@@ -1,4 +1,7 @@
 const express = require('express');
+const advancedResults = require('../middlewares/advancedResults');
+const { protect, authorize } = require('../middlewares/auth');
+const { User } = require('../models/User');
 const {
 	getUsers,
 	getUser,
@@ -8,14 +11,18 @@ const {
 	deleteUsers,
 } = require('../controllers/user');
 
-const { User } = require('../models/User');
-
 const router = express.Router();
 
-const advancedResults = require('../middlewares/advancedResults');
-const { protect, authorize } = require('../middlewares/auth');
+// Include other ressource routers
+const quizRouter = require('./quiz');
+const submissionRouter = require('./submissions');
 
 router.use(protect);
+
+// Re-route into other ressource routers
+router.use('/:id/quiz', quizRouter);
+router.use('/:userId/submissions', submissionRouter);
+
 router.use(authorize('admin'));
 
 router.route('/').get(advancedResults(User), getUsers).patch(updateUsers).delete(deleteUsers);