+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-# 1.0.0 (2020-12-21)
+# 1.0.0 (2020-12-27)
 
 
 ### Features
 
+- Feat: Add Auth, User, Token Routes ([aa9bf61](https://github.com/LouaieRbiha/Quiz_server/commit/aa9bf61561951d10d91fb9f222da332067ffc37f))
 - Feat: Add Token based Auth(Google, Linkedin, Local) ([e2712bb](https://github.com/LouaieRbiha/Quiz_server/commit/e2712bb2b5f57e1c8371a74d6b7ad9f8590c766e))
 - Feat: Add User & Test Models ([a9c28a2](https://github.com/LouaieRbiha/Quiz_server/commit/a9c28a2126865741834be33ba7c275383df56e86))
 
@@ -14,6 +15,7 @@
 
 ### Miscellaneous
 
+- ✨Feat: Add User, Token & Auth Routes ([de28387](https://github.com/LouaieRbiha/Quiz_server/commit/de283871f545d4ea00b1fe14e296ba50cfa3152a))
 - Config changes & redis cache ([cb1f3b4](https://github.com/LouaieRbiha/Quiz_server/commit/cb1f3b404000fdee5352c8922f1f080041fe3caf))
 - Fix package.json start command ([6d06978](https://github.com/LouaieRbiha/Quiz_server/commit/6d06978369685e231d03143ad60237abcaeabb6e))
 - Fix pm2 watch files ([7b85a59](https://github.com/LouaieRbiha/Quiz_server/commit/7b85a5931471fa728be13bc0b970b307938ce24a))